Big restructure
- Move scripts into each workspace
- Rename @kenrick95/c4-core to @kenrick95/c4
- Update folder scructure
