+++ browser/src/board/utils.ts
