+++ browser/src/game/game-online-2p.ts
@@ -1,13 +1,13 @@
 import { Board } from '../board'
-import { BoardBase, BoardPiece } from '@kenrick95/c4-core/board'
+import { BoardBase, BoardPiece } from '@kenrick95/c4/src/board'
 import {
   GameBase,
   MESSAGE_TYPE,
   constructMessage,
   parseMessage,
-} from '@kenrick95/c4-core/game'
-import { Player, PlayerHuman, PlayerShadow } from '@kenrick95/c4-core/player'
-import { Utils } from '@kenrick95/c4-core/utils'
+} from '@kenrick95/c4/src/game'
+import { Player, PlayerHuman, PlayerShadow } from '@kenrick95/c4/src/player'
+import { Utils } from '@kenrick95/c4/src/utils'
 
 enum GAME_MODE {
   FIRST = BoardPiece.PLAYER_1,
