+++ browser/src/game/index.ts
