+++ browser/src/index.html
@@ -66,6 +66,6 @@
       By <a href="http://kenrick95.org">Kenrick95</a>. Fork this project at
       <a href="https://github.com/kenrick95/c4">GitHub</a>.
     </footer>
-    <script src="./browser/app.ts"></script>
+    <script src="./app.ts"></script>
   </body>
 </html>
