+++ browser/tsconfig.json
@@ -12,6 +12,6 @@
     "target": "es5",
     "lib": ["es2015", "dom", "dom.iterable"]
   },
-  "include": ["app.ts", "board/", "game/", "@kenrick95/c4-core"],
+  "include": ["src/", "@kenrick95/c4"],
   "exclude": ["node_modules", "**/*.spec.ts"]
 }
