+++ core/package.json
@@ -1,5 +1,44 @@
 {
-  "name": "@kenrick95/c4-core",
-  "version": "1.0.0",
-  "main": "./lib.ts"
+  "name": "@kenrick95/c4",
+  "version": "4.0.0",
+  "description": "Building blocks for a Connect Four game",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/kenrick95/c4.git"
+  },
+  "author": {
+    "name": "Kenrick",
+    "email": "kenrick95@gmail.com",
+    "url": "https://kenrick95.org"
+  },
+  "bugs": {
+    "url": "https://github.com/kenrick95/c4/issues"
+  },
+  "homepage": "https://github.com/kenrick95/c4",
+  "license": "MIT",
+  "keywords": [
+    "c4",
+    "connect-four",
+    "game"
+  ],
+  "main": "./dist/lib.ts",
+  "source": "./src/lib.ts",
+  "files": [
+    "src/",
+    "dist/"
+  ],
+  "scripts": {
+    "build": "yarn clean && parcel build ./src/lib.ts",
+    "clean": "shx rm -rf dist/",
+    "prettier": "prettier --write \"./**/*.?(ts|html|scss|css)\""
+  },
+  "devDependencies": {
+    "parcel-bundler": "^1.12.4",
+    "prettier": "^2.0.4",
+    "shx": "^0.3.2",
+    "typescript": "^3.8.3"
+  },
+  "resolutions": {
+    "minimist": "^1.2.2"
+  }
 }
