+++ core/src/game/index.ts
