+++ core/src/player/index.ts
