+++ core/src/player/player-ai.ts
