+++ core/tsconfig.json
@@ -12,6 +12,6 @@
     "target": "es5",
     "lib": ["es2015", "dom", "dom.iterable"]
   },
-  "include": ["lib.ts", "utils.ts", "./**/*.ts"],
+  "include": ["src/"],
   "exclude": ["node_modules", "**/*.spec.ts"]
 }
