+++ server/package.json
@@ -1,21 +1,52 @@
 {
   "name": "@kenrick95/c4-server",
   "version": "1.0.0",
+  "description": "Server part of Connect Four game to handle player-to-player communication in browser",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/kenrick95/c4.git"
+  },
+  "author": {
+    "name": "Kenrick",
+    "email": "kenrick95@gmail.com",
+    "url": "https://kenrick95.org"
+  },
+  "bugs": {
+    "url": "https://github.com/kenrick95/c4/issues"
+  },
+  "homepage": "https://github.com/kenrick95/c4",
+  "license": "MIT",
+  "keywords": [
+    "c4",
+    "connect-four",
+    "game"
+  ],
+  "files": [
+    "src/"
+  ],
+  "source": "./src/index.ts",
   "scripts": {
     "start": "ts-node index.ts",
-    "start-dev": "nodemon --watch src --ignore 'src/**/*.spec.ts' --exec ts-node src/index.ts"
+    "start-dev": "nodemon --watch src --ignore 'src/**/*.spec.ts' --exec ts-node src/index.ts",
+    "build": "echo 'OK'",
+    "prettier": "prettier --write \"./**/*.?(ts|html|scss|css)\""
   },
   "devDependencies": {
     "@types/node": "^13.11.1",
     "@types/uuid": "^7.0.2",
     "@types/ws": "^7.2.3",
-    "nodemon": "^2.0.3"
+    "nodemon": "^2.0.3",
+    "ts-node": "^8.8.2",
+    "typescript": "^3.8.3"
   },
   "dependencies": {
-    "@kenrick95/c4-core": "1.0.0",
+    "@kenrick95/c4": "4.0.0",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "uuid": "^7.0.3",
     "ws": "^7.2.3"
+  },
+  "resolutions": {
+    "minimist": "^1.2.2"
   }
 }
