+++ server/src/game/board.ts
@@ -1,3 +1,3 @@
-import { BoardBase } from '@kenrick95/c4-core'
+import { BoardBase } from '@kenrick95/c4/src/board'
 
 export class ServerBoard extends BoardBase {}
