+++ server/src/game/game.ts
@@ -1,14 +1,18 @@
 import { ServerBoard } from './board'
-import { GameBase } from '@kenrick95/c4-core/game'
+import { GameBase } from '@kenrick95/c4/src/game'
 import { ServerPlayer } from './player'
 import { MatchId } from '../types'
-import { BoardPiece } from '@kenrick95/c4-core'
+import { BoardPiece } from '@kenrick95/c4/src/board'
 import { store } from '..'
 import { gameEnded } from '../actions'
 
 export class ServerGame extends GameBase<ServerPlayer> {
   matchId: MatchId
-  constructor(players: Array<ServerPlayer>, board: ServerBoard, matchId: MatchId) {
+  constructor(
+    players: Array<ServerPlayer>,
+    board: ServerBoard,
+    matchId: MatchId
+  ) {
     super(players, board)
     this.matchId = matchId
   }
