+++ server/src/game/player.ts
@@ -1,4 +1,5 @@
-import { PlayerShadow, BoardPiece } from '@kenrick95/c4-core'
+import { BoardPiece } from '@kenrick95/c4/src/board'
+import { PlayerShadow } from '@kenrick95/c4/src/player'
 import { PlayerId } from '../types'
 
 export class ServerPlayer extends PlayerShadow {
